test(backend): move test data loader in main instead of test to allow test data loading in staging profile for staging workflow
